use double for temperature

+++ src/BleSensorListener.h
@@ -34,7 +34,7 @@ public:
     uint8_t broadcastFormat(){ return _broadcastFormat;}
     
     //static int scanForDevice(BTHomeDevicdFoundFunc foundCb);
-    static bool isBleSensorDevice(const NimBLEAdvertisedDevice* device,BleSensorType& type,float& temp,uint8_t& humidity);
+    static bool isBleSensorDevice(const NimBLEAdvertisedDevice* device,BleSensorType& type,double& temp,uint8_t& humidity);
     static BleSensorListener* findBleSensor(const uint8_t mac[6]);
     static BleSensorListener* getBleSensor(const uint8_t mac[6],uint8_t format);
     static void releaseSensor(BleSensorListener* sensor);
